requirements.txt
google-cloud-firestore
starlingbank==2.0